Trait srml_authorship::EventHandler

pub trait EventHandler<Author, BlockNumber> {
    fn note_author(author: Author);

    fn note_uncle(author: Author, age: BlockNumber);
}

An event handler for the authorship module. There is a dummy implementation for (), which does nothing.

Required methods

fn note_author(author: Author)

Note that the given account ID is the author of the current block.

fn note_uncle(author: Author, age: BlockNumber)

Note that the given account ID authored the given uncle, and how many blocks older than the current block it is (age >= 0, so siblings are allowed)

Loading content...